+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Change Log
 
+## [Unreleased]
+- added a wrapper element around the frame, its sizes are taken from config.width and config.height
+
 ## 2.1.0
 - @onlyoffice/docspace-sdk-js:2.0.0